refactor: relocate dedicated components of `Skills`
