+++ src/features/skills/Skills.tsx
@@ -1,4 +1,4 @@
-import SkillsContainer from "@/features/skills/SkillsContainer";
+import SkillsContainer from "@/features/skills/components/SkillsContainer";
 import { JavaIcon } from "@/utils/icons";
 import React from "react";
 import {
